+++ src/index.ts
@@ -1,20 +1,11 @@
 export * from './types';
 
-import { Canvas, CanvasProps } from './canvas';
-import { ConsumerProps, createContext, ProviderProps } from './create-context';
+import { Canvas } from './canvas';
+import { createContext } from './create-context';
 import { createElement } from './create-element';
 import { render } from './render';
 
-export {
-  Canvas,
-  createContext,
-  createElement,
-  render,
-  // Types
-  CanvasProps,
-  ConsumerProps,
-  ProviderProps,
-};
+export { Canvas, createContext, createElement, render };
 
 export default {
   Canvas,
